Initial drawable code
The drawable code is in the repository along with the sphere class for
testing and the corresponding .fx file. The Luna code looks like it
actually has the Effects library (which appears to not exist on lab
machines) so I think it would be better to switch the class back using
effects, rather than individual vertex and pixel shaders, as that will
work much better.